SimplyAgree: Flexible and Robust Agreement and Reliability Analyses

Reliability and agreement analyses often have limited software support. Therefore, this package was created to make agreement and reliability analyses easier for the average researcher. The functions within this package include simple tests of agreement, agreement analysis for nested and replicate data, and provide robust analyses of reliability. In addition, this package contains a set of functions to help when planning studies looking to assess measurement agreement.

See also

Author

Maintainer: Aaron Caldwell arcaldwell49@gmail.com