Tests for Absolute Agreement with Nested Data

agree_nest produces an absolute agreement analysis for data where there is multiple observations per subject but the mean varies within subjects as described by Zou (2013). Output mirrors that of agree_test but CCC is calculated via U-statistics.

agree_nest(x, y, id, data, delta, agree.level = 0.95, conf.level = 0.95)

Arguments

x
Name of column with first measurement

y
Name of other column with first measurement

id
Column with subject identifier

data
Data frame with all data

delta
The threshold below which methods agree/can be considered equivalent, can be in any units. Equivalence Bound for Agreement.

agree.level
the agreement level required. Default is 95%. The proportion of data that should lie between the thresholds, for 95% limits of agreement this should be 0.95.

conf.level
the confidence level required. Default is 95%.

Value

Returns single list with the results of the agreement analysis.

"loa"

a data frame of the limits of agreement including the average difference between the two sets of measurements, the standard deviation of the difference between the two sets of measurements and the lower and upper confidence limits of the difference between the two sets of measurements.

"h0_test"

Decision from hypothesis test.

"identity.plot"

Plot of x and y with a line of identity with a linear regression line

"bland_alt.plot"

Simple Bland-Altman plot. Red line are the upper and lower bounds for shieh test; grey box is the acceptable limits (delta). If the red lines are within the grey box then the shieh test should indicate 'reject h0', or to reject the null hypothesis that this not acceptable agreement between x & y.

"ccc.xy"

Lin's concordance correlation coefficient and confidence intervals using U-statistics. Warning: if underlying value varies this estimate will be inaccurate.

"conf.level"

Returned as input.

"agree.level"

Returned as input.

References

Zou, G. Y. (2013). Confidence interval estimation for the Bland–Altman limits of agreement with multiple observations per individual. Statistical methods in medical research, 22(6), 630-642.

King, TS and Chinchilli, VM. (2001). A generalized concordance correlation coefficient for continuous and categorical data. Statistics in Medicine, 20, 2131:2147.

King, TS; Chinchilli, VM; Carrasco, JL. (2007). A repeated measures concordance correlation coefficient. Statistics in Medicine, 26, 3095:3113.

Carrasco, JL; Phillips, BR; Puig-Martinez, J; King, TS; Chinchilli, VM. (2013). Estimation of the concordance correlation coefficient for repeated measures using SAS and R. Computer Methods and Programs in Biomedicine, 109, 293-304.

Examples

data('reps')
agree_nest(x = "x", y = "y", id = "id", data = reps, delta = 2)

#> Limit of Agreement = 95%
#> alpha = 0.05 | 95 % Confidence Interval
#> Nested Data Points (true value may vary)
#> 
#> Hypothesis Test: don't reject h0
#> 
#> ###- Bland-Altman Limits of Agreement (LoA) -###
#> Mean Bias: 0.7100833 [-0.6824145, 2.102581]
#> Lower LoA: -2.153994 [-9.822981, -0.3468361]
#> Upper LoA: 3.574161 [1.767003, 11.24315]
#> 
#> ###- Concordance Correlation Coefficient (CCC) -###
#> CCC: 0.4791, 95% C.I. [0.2429, 0.6616]
#> *Estimated via U-statistics; may be biased