Reliability Statistics

reli_stats produces reliability statistics desccribed by Weir (2005). This includes intraclass correlation coefficients, the coefficient of variation, and the standard error of meassurement.

reli_stats(
  measure,
  item,
  id,
  data,
  wide = FALSE,
  col.names = NULL,
  conf.level = 0.95
)

Arguments

measure

Name of column containing the measurement of interest

item

Name of column containing the items. If this is a test-retest reliability study then this would indicate the time point (e.g., time1,time2, time3, etc.)

id

Column with subject identifier

data

Data frame with all data

wide

Logical value (TRUE or FALSE) indicating if data is in a "wide" format. Default is TRUE.

col.names

If wide is equal to TRUE then col.names is a list of the column names containing the measurements for reliability analysis.

conf.level

the confidence level required. Default is 95%.

Value

Returns single list with the results of the agreement analysis.

"icc"

Table of ICC results

"lmer"

Linear mixed model from lme4

"anova"

Analysis of Variance table

"var_comp"

Table of Variance Components

"n.id"

Number of subjects/participants

"n.items"

Number of items/time points

"cv"

Coefficient of Variation

"SEM"

Standard Error of Measurement

"SEE"

Standard Error of the Estimate

"SEP"

Standard Error of Predicitions

"plot.reliability"

Plot of data points within subjects across items

References

Weir, J. P. (2005). Quantifying test-retest reliability using the intraclass correlation coefficient and the SEM. The Journal of Strength & Conditioning Research, 19(1), 231-240.

Examples

data('reps')
reli_stats(data = reps, wide = TRUE, col.names = c("x","y"))
#> 
#> Coefficient of Variation (%):  12.35
#> Standard Error of Measurement (SEM):  0.8634
#> Standard Error of the Estimate (SEE):  0.9145
#> Standard Error of Prediction (SEP):  1.5457
#> 
#> Intraclass Correlation Coefficients
#>            model         measures  type    icc lower.ci upper.ci
#> 1 one-way random        Agreement  ICC1 0.4963   0.1632   0.7299
#> 2 two-way random        Agreement  ICC2 0.5097   0.1911   0.7355
#> 3  two-way fixed      Consistency  ICC3 0.5384   0.2117   0.7569
#> 4 one-way random   Avg. Agreement ICC1k 0.6634   0.2806   0.8438
#> 5 two-way random   Avg. Agreement ICC2k 0.6753   0.3209   0.8476
#> 6  two-way fixed Avg. Consistency ICC3k 0.7000   0.3495   0.8616
#>