Tests for Absolute Agreement

Tests for Absolute Agreement

agree_test(x, y, delta, conf.level = 0.95, agree.level = 0.95)

Arguments

x
Vector with first measurement

y
Vector with second measurement

delta
The threshold below which methods agree/can be considered equivalent, can be in any units. Equivalence Bound for Agreement.

conf.level
the confidence level required. Default is 95%.

agree.level
the agreement level required. Default is 95%. The proportion of data that should lie between the thresholds, for 95% limits of agreement this should be 0.95.

Value

Returns single list with the results of the agreement analysis.

"shieh_test"

The TOST hypothesis test as described by Shieh.

"ccc.xy"

Lin's concordance correlation coefficient and confidence intervals.

"s.shift"

Scale shift from x to y.

"l.shift"

Location shift from x to y.

"bias"

a bias correction factor that measures how far the best-fit line deviates from a line at 45 degrees. No deviation from the 45 degree line occurs when bias = 1. See Lin 1989, page 258.

"loa"

Data frame containing the limits of agreement calculations

"h0_test"

Decision from hypothesis test.

"identity.plot"

Plot of x and y with a line of identity with a linear regression line

"bland_alt.plot"

Simple Bland-Altman plot. Red line are the upper and lower bounds for shieh test; grey box is the acceptable limits (delta). If the red lines are within the grey box then the shieh test should indicate 'reject h0', or to reject the null hypothesis that this not acceptable agreement between x & y.

References

Gwowen Shieh (2019): Assessing Agreement Between Two Methods of Quantitative Measurements: Exact Test Procedure and Sample Size Calculation, Statistics in Biopharmaceutical Research, <https://doi.org/10.1080/19466315.2019.1677495>

Examples

#to be added