Reliability Statistics

Reliability Statistics

reli_stats(
  measure,
  item,
  id,
  data,
  wide = FALSE,
  col.names = NULL,
  conf.level = 0.95
)

Arguments

measure
Name of column containing the measurement of interest

item
Name of column containing the items. If this is a test-retest reliability study then this would indicate the time point (e.g., time1,time2, time3, etc.)

id
Column with subject identifier

data
Data frame with all data

wide
Logical value (TRUE or FALSE) indicating if data is in a "wide" format. Default is TRUE.

col.names
If wide is equal to TRUE then col.names is a list of the column names containing the measurements for reliability analysis.

conf.level
the confidence level required. Default is 95%.

Value

Returns single list with the results of the agreement analysis.

"icc"

Table of ICC results

"lmer"

Linear mixed model from lme4

"anova"

Analysis of Variance table

"var_comp"

Table of Variance Components

"n.id"

Number of subjects/participants

"n.items"

Number of items/time points

"cv"

Coefficient of Variation

"SEM"

Standard Error of Measurement

"SEE"

Standard Error of the Estimate

"SEP"

Standard Error of Predicitions

"plot.reliability"

Plot of data points within subjects across items

References

Weir, J. P. (2005). Quantifying test-retest reliability using the intraclass correlation coefficient and the SEM. The Journal of Strength & Conditioning Research, 19(1), 231-240.

Examples

#to be added